+++ tsdb/head_wal.go
@@ -74,6 +74,11 @@ func (h *Head) loadWAL(r *wlog.Reader, syms *labels.SymbolTable, multiRef map[ch
 		decodeErr, seriesCreationErr error
 	)
 
+	_, last, err := wlog.Segments(h.wal.Dir())
+	if err != nil {
+		return fmt.Errorf("failed to get last segment to set WAL expiry for duplicate series: %w", err)
+	}
+
 	defer func() {
 		// For CorruptionErr ensure to terminate all workers before exiting.
 		_, ok := err.(*wlog.CorruptionErr)
@@ -237,6 +242,8 @@ Outer:
 				}
 				if !created {
 					multiRef[walSeries.Ref] = mSeries.ref
+					// Set the WAL expiry for the duplicate series, so it is kept in subsequent WAL checkpoints.
+					h.setWALExpiry(walSeries.Ref, last)
 				}
 
 				idx := uint64(mSeries.ref) % uint64(concurrency)